Safari-609.1.11.1/LayoutTests/crypto/subtle/aes-import-raw-key-export-raw-key-expected.txt

Test importing a raw AES-CBC key with legnth 128 and then export it in raw format

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Importing a key...
Exporting a key...
PASS exportedRawKey is rawKey
PASS successfullyParsed is true

TEST COMPLETE
